pyproject.toml
[tool.poetry]
name = "adventuretrackerapi"
version = "0.1.0"
description = ""
authors = ["Jakub Dubec <jakub.dubec@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
django = "==3.0.*"
python-dotenv = "^0.12.0"
sentry-sdk = "^0.14.3"
psycopg2-binary = "^2.8.4"
django-enum-choices = "^2.1.2"
pillow = "^7.0.0"
argon2-cffi = "^19.2.0"
django_api_forms = "^0.7.0"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
fabric = "^2.5.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
